+++ src/Braintree/PaymentMethodOptionsRequest.cs
@@ -4,13 +4,14 @@ namespace Braintree
 {
     public class PaymentMethodOptionsRequest : Request
     {
-        public bool? MakeDefault { get; set; }
-        public bool? VerifyCard { get; set; }
         public bool? FailOnDuplicatePaymentMethod { get; set; }
-        public string VerificationMerchantAccountId { get; set; }
+        public bool? MakeDefault { get; set; }
+        public bool? SkipAdvancedFraudChecking { get; set; }
+        public bool? VerifyCard { get; set; }
+        public string VerificationAccountType { get; set; } // NEXT_MAJOR_VERSION - This should be an enum with [credit, debit]
         public string VerificationAmount { get; set; }
-        public string VerificationAccountType { get; set; }
         public string VerificationCurrencyIsoCode { get; set; }
+        public string VerificationMerchantAccountId { get; set; }
         public PaymentMethodOptionsPayPalRequest OptionsPayPal { get; set; }
         public UsBankAccountVerificationMethod? UsBankAccountVerificationMethod { get; set; }
 
@@ -27,15 +28,16 @@ namespace Braintree
         protected virtual RequestBuilder BuildRequest(string root)
         {
             return new RequestBuilder(root).
-                AddElement("make-default", MakeDefault).
-                AddElement("verification-merchant-account-id", VerificationMerchantAccountId).
-                AddElement("verification-account-type", VerificationAccountType).
-                AddElement("verification-currency-iso-code", VerificationCurrencyIsoCode).
-                AddElement("verify-card", VerifyCard).
-                AddElement("us-bank-account-verification-method", UsBankAccountVerificationMethod.GetDescription()).
-                AddElement("verification-amount", VerificationAmount).
                 AddElement("fail-on-duplicate-payment-method", FailOnDuplicatePaymentMethod).
-                AddElement("paypal", OptionsPayPal);
+                AddElement("make-default", MakeDefault).
+                AddElement("paypal", OptionsPayPal).
+                AddElement("skip-advanced-fraud-checking", SkipAdvancedFraudChecking).
+                AddElement("us-bank-account-verification-method", UsBankAccountVerificationMethod.GetDescription()).
+                AddElement("verification-account-type", VerificationAccountType).
+                AddElement("verification-amount", VerificationAmount).
+                AddElement("verification-currency-iso-code", VerificationCurrencyIsoCode).
+                AddElement("verification-merchant-account-id", VerificationMerchantAccountId).
+                AddElement("verify-card", VerifyCard);
         }
     }
 }